Create base class to define to/from flat and misc methods
